feature(service describe): Output of service details
As discussed in #48 this `kn service describe` as changed in this commit
mimics `kubectl describe <sth>` as that it output detail information
in human readable output.

Summary information of a list of services or a single services as well
as exporting the service in various formats (JSON, YAML) is reserved to
`kn service get`

For this reason, the generic Printer handling from cli-runtime has been
removed as `kn service describe` only outputs human readable format
(that is btw also true for `kubectl` describe which does not allow for
`-o` kind of options).

This command shows information about the serivce itself, but also a summary about the associated revisions.

"kn service describe" knows three modes, which can be combined:

`--all`    : Print more information. By default only are shorter summary is shown
`--color`  : Use a colorful output (but only when on a tty)
`--hipster`: Experimental output mode for a very compact representation using emojis

Of course there can be more and other information to be shown. Let's discuss and add this (and maybe remove the dense format)