Adds a Ginkgo-style unit test to `kn version` command
1. adds suites for each package where commands reside
2. adds version_test.go to completely unit test version
   as an example of what can be done

NOTE also that all current Go tests are still compatible
and runnable. So if we decide to start using this fashion
for UTs then we can gradually do so. I'll make some comments
about some advantages of Ginkgo/Gomega in issue #161 which
is the one we are using to track discussion.