Reorg OWNERS and introduced OWNERS_ALIASES
This update is for aligning with the other Knative projects so
that a group like the TOC list can be updated with less efforts
in the future.

Also took the occasion to clean up a bit, so removed
@cppforlife and @sixolet as they are not active in the project since
at least one year. Happy to re-add both accounts again in the future.

Thanks a lot Naomi and Dmitriy for all your contribution and kicking off the client !